Refactor db module
