+++ tests/conftest.py
@@ -16,10 +16,11 @@ from crawler.constants import (
     FIELD_FILTERED_POSITIVE,
     MLWH_TABLE_NAME,
 )
-from crawler.db import create_mongo_client, create_mysql_connection, get_mongo_collection, get_mongo_db
+from crawler.db.mongo import create_mongo_client, get_mongo_collection, get_mongo_db
+from crawler.db.mysql import create_mysql_connection
 from crawler.file_processing import Centre, CentreFile
 from crawler.helpers.general_helpers import get_config
-from tests.testing_objects import (
+from tests.data.testing_objects import (
     EVENT_WH_DATA,
     FILTERED_POSITIVE_TESTING_SAMPLES,
     MLWH_SAMPLE_LIGHTHOUSE_SAMPLE,
